src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
bitarray==3.4.2
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
contourpy==1.3.0
cycler==0.12.1
cytoolz==1.0.1
eth-abi==5.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
fonttools==4.58.1
frozenlist==1.6.2
hexbytes==1.3.1
idna==3.10
importlib-resources==6.5.2
joblib==1.5.1
kiwisolver==1.4.7
matplotlib==3.9.4
multidict==6.4.4
numpy==2.0.2
packaging==25.0
pandas==2.3.0
parsimonious==0.10.0
pillow==11.2.1
propcache==0.3.1
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-core==2.33.2
pymdptoolbox==4.0b3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
threadpoolctl==3.6.0
toolz==1.0.0
types-requests==2.32.0.20250602
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
web3==7.12.0
websockets==15.0.1
yarl==1.20.0
z3-solver==4.15.0.0
zipp==3.22.0
